+++ changelog.md
@@ -5,6 +5,8 @@
 
 * [fixed] Download ZIP with media files in production
 * [changed]  Length of individuality prolonged to 100 characters
+* [added] Ordering in Locailities
+* [changed] Generic Locality List view used
 
 ## 01-2025
 * [added] Added carousel for media files in the detail view